+++ src/Lykke.Job.TxDetector/Events/CashInOperationCreatedEvent.cs
@@ -1,10 +1,10 @@
-﻿using Lykke.Job.TxDetector.Sagas.Models;
+﻿using Lykke.Job.TxDetector.Models;
 using ProtoBuf;
 
-namespace Lykke.Job.TxDetector.Sagas.Events
+namespace Lykke.Job.TxDetector.Events
 {
     [ProtoContract]
-    public class CashInOutOperationRegisteredEvent
+    public class CashInOperationCreatedEvent
     {
         [ProtoMember(1)]
         public Transaction Transaction { get; set; }
@@ -12,7 +12,5 @@ namespace Lykke.Job.TxDetector.Sagas.Events
         public Asset Asset { get; set; }
         [ProtoMember(3)]
         public double Amount { get; set; }
-        [ProtoMember(4)]
-        public string CommandId { get; set; }
     }
 }
