+++ src/Lykke.Job.TxDetector/Handlers/TransferHandler.cs
@@ -3,10 +3,11 @@ using System.Threading.Tasks;
 using Common;
 using Common.Log;
 using JetBrains.Annotations;
+using Lykke.Job.TxDetector.Commands;
 using Lykke.Job.TxDetector.Core.Domain.PaymentSystems;
-using Lykke.Job.TxDetector.Sagas.Commands;
+using Lykke.Job.TxDetector.Sagas;
 
-namespace Lykke.Job.TxDetector.Sagas.Handlers
+namespace Lykke.Job.TxDetector.Handlers
 {
     public class TransferHandler
     {
